requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
deepface==0.0.75
docutils==0.19
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
mediapipe==0.9.0.1
mtcnn==0.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
